requirements.txt
Flask>=2.0.0
flask-cors>=4.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
psutil>=5.8.0
SQLAlchemy>=1.4.0
graphviz>=0.20.0